pyproject.toml
[project]
name = "goit-pythonweb-hw-12"
version = "1.0.0"
description = ""
authors = [
    {name = "Max Novik",email = "novik.kor@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "alembic (>=1.16.5,<2.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "fastapi-mail (>=1.4.1,<2.0.0)",
    "jinja2 (>=3.1.4,<4.0.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "fakeredis (>=2.31.1,<3.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
mypy = "^1.17.1"
black = "^25.1.0"
types-passlib = "^1.7.7.20250602"
types-python-jose = "^3.5.0.20250531"
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
pytest-asyncio = "^1.2.0"
fakeredis = "^2.31.1"
sphinx = "^8.2.3"
aiosqlite = "^0.21.0"
pytest-mock = "^3.15.1"
sphinx-markdown-builder = "^0.6.8"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
